+++ src/headless_re_mcp/backends/frida/__init__.py
@@ -1,3 +1,7 @@
-from headless_re_mcp.backends.frida.client import FridaClient, FridaError
+from headless_re_mcp.backends.frida.client import (
+    HOOK_TEMPLATE_NAMES,
+    FridaClient,
+    FridaError,
+)
 
-__all__ = ['FridaClient', 'FridaError']
+__all__ = ['HOOK_TEMPLATE_NAMES', 'FridaClient', 'FridaError']